@@ -10,7 +10,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.12.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -18,10 +18,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  intl: ^0.16.1
-  fluttertoast: ^3.0.3
-  english_words: ^3.1.0
+  cupertino_icons: ^1.0.2
+  intl: ^0.17.0
+  fluttertoast: ^8.0.3
+  english_words: ^4.0.0
 
 dev_dependencies:
   flutter_test:
